gpu: nova-core: mm: Add TLB flush support
Add TLB (Translation Lookaside Buffer) flush support for GPU MMU.

After modifying page table entries, the GPU's TLB must be invalidated
to ensure the new mappings take effect. The Tlb struct provides flush
functionality through BAR0 registers.

The flush operation writes the page directory base address and triggers
an invalidation, polling for completion with a 2 second timeout matching
the Nouveau driver.

Cc: Nikola Djukic <ndjukic@nvidia.com>

+++ drivers/gpu/nova-core/fsp.rs
@@ -485,7 +485,7 @@ impl Fsp {
         let nvdm = mctp::NvdmHeader::from_raw(nvdm_header);
         if nvdm.msg_type() != mctp::MSG_TYPE_VENDOR_PCI
             || nvdm.vendor_id() != mctp::VENDOR_ID_NV
-            || nvdm.nvdm_type() != mctp::nvdm_type::FSP_RESPONSE
+            || nvdm.nvdm_type() != u32::from(mctp::nvdm_type::FSP_RESPONSE)
         {
             dev_err!(
                 dev,
